Move ViolationGenerationEngine
